requirements.txt
openai==1.66.2
langchain==0.3.20
langchain-openai==0.3.8
python-dotenv==1.0.1
# faiss-cpu==1.10.0
langchain-community==0.3.19
# pypdf==5.3.1